Move lexical model steps
